Bit of an experiment with one possible trajectory animation
Not sure if I like this. What's good about it is that it's quite a
simulation-oriented approach. Every n milliseconds, the simulation ticks and
moves the missile forward by however many metres it should have moved since the
last tick.

But this is problematic because of all the straight-line movement which looks
flickery as it gets evened out over the course of the flight. Maybe the
interpolation approach is the way to go after all.